requirements.txt
Flask==2.2.5
tensorflow==2.12.0
pillow==10.2.0
numpy==1.24.3
gunicorn==21.2.0